Fix login with json, comment swagger, to extract jwt
