+++ patient-portal/src/main/java/org/isf/patientportal/config/SwaggerConfig.java.old
