+++ patient-portal/src/main/java/org/isf/patientportal/security/CustomAuthenticationManager.java
@@ -19,7 +19,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  */
-package org.isf.security;
+package org.isf.patientportal.security;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
